% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfolds.R
\name{kfolds}
\alias{kfolds}
\title{k-fold cross-validation indices}
\usage{
kfolds(n, K, seed = NULL)
}
\arguments{
\item{n}{number of data points (positive integer)}

\item{K}{number of folds (integer K >= 2)}

\item{seed}{sets random seed for reproducibility (positive integer)}
}
\value{
A list of length two, each entry itself being a list of
length \code{K} containing the training (validation) indices.
}
\description{
Produces k-fold partitions of training and validation indices for
a dataset.
}
\note{
See also \code{\link{tsmvr_cv}}.
}
