% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_hub.R
\name{precision_hub}
\alias{precision_hub}
\title{Precision matrix of scale free network}
\usage{
precision_hub(q, g, edge_val = 0.3, gamma = 1e-06)
}
\arguments{
\item{q}{square dimension of precision matrix (positive numeric integer)}

\item{g}{number of hubs, (numeric integer st. 1 <= g <= q)}

\item{edge_val}{values of hub network edges (numeric st. -1/sqrt(2)+gamma <= edge_val <= 1/sqrt(2)-gamma)}

\item{gamma}{numeric value that keeps edge_val bounded st. the returned precision matrix is PSD}
}
\value{
Returns an Hub graph precision matrix
}
\description{
Constructs a Hub network precision matrix
}
\references{
\insertRef{chen2016high}{tsmvr}
}
