% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parameters.R
\name{set_parameters}
\alias{set_parameters}
\title{Set parameters for tsmvr_solve}
\usage{
set_parameters(B_type = "ls", Omega_type = "ls", eta1 = 0.01,
  eta2 = 0.01, lam1 = 0.1, lam2 = 0.1, del1 = 0.001,
  del2 = 0.001, rho1 = 0.001, rho2 = 0.001, beta1 = 0.5,
  beta2 = 0.5, qmax1 = 128, qmax2 = 128, eps1 = 1e-04,
  eps2 = Inf, k = 5, reps = 10, max_iter = 2000, skip = 10,
  quiet = F, suppress = F, seed = NULL)
}
\arguments{
\item{B_type}{type of descent for regression steps (string: 'gd' or 'ls')}

\item{Omega_type}{(string: 'gd', 'ls' or 'min')}

\item{eta1}{B-step learning rate (positive numeric)}

\item{eta2}{Omega-step learning rate (positive numeric)}

\item{lam1}{B-step learning rate (positive numeric)}

\item{lam2}{Omega-step learning rate (positive numeric)}

\item{del1}{B initialization matrix inversion fudge factor (positive numeric)}

\item{del2}{Omega initialization matrix inversion fudge factor (positive numeric)}

\item{rho1}{B-step linesearch convergence parameter (positive numeric)}

\item{rho2}{Omega-step linesearch convergence parameter (positive numeric)}

\item{beta1}{B-step linesearch stepsize shrinking paramter (positive numeric)}

\item{beta2}{Omega-step linesearch stepsize shrinking paramter (positive numeric)}

\item{qmax1}{B-step linesearch maximum number of iterations (positive integer valued numeric)}

\item{qmax2}{Omega-step linesearch maximum number of iterations (positive integer valued numeric)}

\item{eps1}{B-step convergence parameter (positive numeric)}

\item{eps2}{Omega-step convergence parameter (positive numeric)}

\item{k}{number of cross-validation folds (positive integer valued numeric greater than 1)}

\item{reps}{number of k-fold cross-validation replicates (positive integer valued numeric)}

\item{max_iter}{maximum number of iterations (positive integer)}

\item{skip}{iteration skip frequency for output to screen (positive integer)}

\item{quiet}{whether or not to operate   (bool)}

\item{suppress}{whether or not to suppress warnings (bool)}

\item{seed}{sets random seed}
}
\value{
Returns a named list of input parameters.

See also \code{\link{tsmvr_solve}},
\code{\link{tsmvr_cv}},
\code{\link{tsmvr_replicate}}, and
\code{\link{tsmvr_gridsearch}}.
}
\description{
Sets paramters for tsmvr_solve and functions that call tsmvr_solve
(tsmvr_cv, tsmvr_replicate, and tsmvr_gridsearch).
}
\references{
\insertRef{chen2016high}{tsmvr}
\insertRef{chen2018covariate}{tsmvr}
}
