% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmvr.R
\name{tsmvr}
\alias{tsmvr}
\title{Truly Spare MultiVariate Regression}
\usage{
tsmvr(X, Y, s1 = NULL, s2 = NULL, s1_vec = NULL, s2_vec = NULL,
  method = c("single", "gs"), pars, quiet = FALSE, seed = NULL)
}
\arguments{
\item{X}{n-by-p predictor matrix}

\item{Y}{n-by-q response matrix}

\item{s1}{sparsity for p-by-q regressor matrix \code{B} (positive integer or NULL)}

\item{s2}{sparsity for q-by-q precision matrix \code{Omega} (positive integer or NULL)}

\item{s1_vec}{\code{s1} values for gridsearch (vector of integer valued numerics or NULL)}

\item{s2_vec}{\code{s2} values for gridsearch (vector of integer valued numerics or NULL)}

\item{pars}{list of algorithm parameters as constructed by the \code{set_parameters} function}

\item{seed}{set random seed (integer or NULL)}
}
\value{
A list containing the \code{mean} and \code{sd} of the
error over the replicates as well as the means and standard
deviations of the errors across each fold.
}
\description{
Calculates sparse solutions to the multivariate regression problem with
correlated errors. Let \eqn{X} be the \eqn{n-by-p} predictor matrix,
\eqn{Y} be the \eqn{n-by-q} response matrix, \eqn{B} be the \eqn{p-by-q}
regression matrix, and \eqn{\Omega} be the \eqn{q-by-q} precision matrix
(inverse of the covariance matrix \eqn{\Sigma}). Let \eqn{s1} be a sparsity
parameter that only allows the top \eqn{s1} values for \eqn{B} in aboslute
value be nonzero, and lets \eqn{s2} be defined similarly for \eqn{\Omega}. Then
this functions calculates the solution to
\deqn{(B_hat,\Omega_hat) = argmin_(B,\Omega)
(1/n*Tr[(Y-XB)^T*\Omega*(Y-XB)] - log|\Omega}|)
such that \eqn{||B||_(1,1) \le s1} and \eqn{||\Omega||_(1,1) \le s2}. Here
\eqn{|| * ||_(1,1)} indicates the 1-1 "norm" that counts the number of nonzero
matrix entries.
}
\note{
See also \code{\link{set_parameters}}.
}
