% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmvr_cv.R
\name{tsmvr_cv}
\alias{tsmvr_cv}
\title{k-fold cross-validation for tsmvr}
\usage{
tsmvr_cv(X, Y, s1, s2, K = 10, b_type = "gd", omega_type = "min",
  eta1 = 0.01, eta2 = 0.01, epsilon = 1e-05, maxiter = 1000,
  quiet = FALSE, seed = NULL)
}
\arguments{
\item{X}{design matrix (n-by-p)}

\item{Y}{response matrix (n-by-q)}

\item{s1}{regressor matrix sparsity (positive integer)}

\item{s2}{covariance matrix sparsity (positive integer)}

\item{K}{number of folds (integer greater than 1)}

\item{b_type}{B-step descent type (string: 'gd')}

\item{omega_type}{Omega-step descent type (string: 'gd' or 'min')}

\item{eta1}{B-step learning rate (positive numeric)}

\item{eta2}{Omega-step learning rate (positive numeric)}

\item{epsilon}{convergence parameter (positive numeric)}

\item{maxiter}{maximum number of allowed iterations (positive integer)}

\item{quiet}{quiet mode (bool)}

\item{seed}{set random seed (integer)}
}
\value{
A list of the mean and standard deviation of the errors
across the \code{K} folds.
}
\description{
Calculates the mean and standard deviation of errors
over a tsmvr k-fold cross-validation experiment.
The error for each base model is the normalized squared
error between the true response and the predicted response
on a given cross-validation set.
}
\note{
See also \code{\link{squared_error}}, \code{\link{kfolds}},
\code{\link{tsmvr_solve}}.
}
