% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmvr_replicate.R
\name{tsmvr_replicate}
\alias{tsmvr_replicate}
\title{Replicated k-fold cross-validation for tsmvr}
\usage{
tsmvr_replicate(X, Y, s1, s2, K = 10, reps = 1, b_type = "gd",
  omega_type = "min", eta1 = 0.01, eta2 = 0.01, epsilon = 1e-05,
  maxiter = 1000, quiet = FALSE, seed = NULL)
}
\arguments{
\item{X}{design matrix (n-by-p)}

\item{Y}{response matrix (n-by-q)}

\item{s1}{regressor matrix sparsity (positive integer)}

\item{s2}{covariance matrix sparsity (positive integer)}

\item{K}{number of folds (integer greater than 1)}

\item{reps}{number of replications (positive integer)}

\item{b_type}{B-step descent type (string: 'gd')}

\item{omega_type}{Omega-step descent type (string: 'gd' or 'min')}

\item{eta1}{B-step learning rate (positive numeric)}

\item{eta2}{Omega-step learning rate (positive numeric)}

\item{epsilon}{convergence parameter (positive numeric)}

\item{maxiter}{maximum number of allowed iterations (positive integer)}

\item{quiet}{quiet mode (bool)}

\item{seed}{set random seed (integer)}
}
\value{
A list containing the \code{mean} and \code{sd} of the
error over the replicates as well as the means and standard
deviations of the errors across each fold.
See also \code{\link{squared_error}}, \code{\link{kfolds}},
\code{\link{tsmvr_cv}}, and \code{\link{tsmvr_solve}}.
}
\description{
Calculates the mean and standard deviation of fold-avereged errors
derived from replcated tsmvr k-fold cross-validation experiments.
The error for each base model is the normalized squared
error between the true response and the predicted response
on a given cross-validation set.
}
