% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_folds.R
\name{k_folds}
\alias{k_folds}
\title{k-fold cross-validation indices}
\usage{
k_folds(n, k, seed = NULL)
}
\arguments{
\item{n}{number of data points (integer valued numeric > 1)}

\item{k}{number of k-folds (integer valued numeric > 1)}

\item{seed}{sets random seed for reproducibility (numeric)}
}
\value{
A list of length two, each entry itself being a list of
length \code{K} containing the training (validation) indices.
}
\description{
Produces k-fold partitions of training and validation indices for
a dataset.
}
\note{
See also \code{\link{tsmvr_cv}}.
}
