% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmvr_cv.R
\name{tsmvr_cv}
\alias{tsmvr_cv}
\title{k-fold cross-validation for tsmvr}
\usage{
tsmvr_cv(X, Y, s1, s2, pars, quiet = F, seed = NULL)
}
\arguments{
\item{X}{design matrix (n-by-p)}

\item{Y}{response matrix (n-by-q)}

\item{s1}{regressor matrix sparsity (positive integer)}

\item{s2}{covariance matrix sparsity (positive integer)}

\item{pars}{list of algorithm parameters; output of \code{set_parameters}}

\item{quiet}{(logical)}

\item{seed}{set random seed (integer)}
}
\value{
A list of the mean and standard deviation of the errors
across the \code{K} folds.
}
\description{
Calculates the mean and standard deviation of errors
over a tsmvr k-fold cross-validation experiment.
The error for each base model is the normalized squared
error between the true response and the predicted response
on a given cross-validation set.
}
\note{
See also \code{\link{k_folds}},
\code{\link{tsmvr}}, and
\code{\link{set_parameters}}.
}
