% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmvr_cv.R
\name{tsmvr_cv}
\alias{tsmvr_cv}
\title{k-fold cross-validation for tsmvr}
\usage{
tsmvr_cv(X, Y, s1, s2, k = 10, B_type = "ls", Omega_type = "ls",
  eta1 = 0.05, eta2 = 0.2, rho1 = 100, rho2 = 100, beta1 = 0.5,
  beta2 = 0.5, epsilon = 0.001, max_iter = 2000, quiet = FALSE,
  suppress = FALSE, seed = NULL)
}
\arguments{
\item{X}{design matrix (n-by-p)}

\item{Y}{response matrix (n-by-q)}

\item{s1}{regressor matrix sparsity (positive integer)}

\item{s2}{covariance matrix sparsity (positive integer)}

\item{k}{number of k-folds (integer greater than 1)}

\item{B_type}{B-step descent type (string: 'gd')}

\item{Omega_type}{Omega-step descent type (string: 'gd' or 'min')}

\item{eta1}{B-step learning rate (positive numeric)}

\item{eta2}{Omega-step learning rate (positive numeric)}

\item{rho1}{B-step linesearch convergence parameter (positive numeric)}

\item{rho2}{Omega-step linesearch convergence parameter  (positive numeric)}

\item{beta1}{B-step linesearch shrinkage parameter (positive numeric)}

\item{beta2}{Omega-step linesearch shrinkage parameter (positive numeric)}

\item{epsilon}{convergence parameter (positive numeric)}

\item{max_iter}{maximum number of allowed iterations (positive integer)}

\item{quiet}{quiet mode (bool)}

\item{suppress}{suppress maximum iteration warning (bool)}

\item{seed}{set random seed (integer)}
}
\value{
A list of the mean and standard deviation of the errors
across the \code{K} folds.
}
\description{
Calculates the mean and standard deviation of errors
over a tsmvr k-fold cross-validation experiment.
The error for each base model is the normalized squared
error between the true response and the predicted response
on a given cross-validation set.
}
\note{
See also \code{\link{squared_error}}, \code{\link{k_folds}},
\code{\link{tsmvr_solve}}.
}
