% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmvr_gridsearch.R
\name{tsmvr_gridsearch}
\alias{tsmvr_gridsearch}
\title{Gridsearch for tsmvr}
\usage{
tsmvr_gridsearch(X, Y, s1_grid, s2_grid, k = 10, reps = 10,
  B_type = "ls", Omega_type = "ls", eta1 = 0.05, eta2 = 0.2,
  rho1 = 100, rho2 = 100, beta1 = 0.5, beta2 = 0.5,
  epsilon = 1e-05, max_iter = 2000, suppress = TRUE, quiet = FALSE,
  seed = NULL)
}
\arguments{
\item{X}{design matrix (n-by-p)}

\item{Y}{response matrix (n-by-q)}

\item{s1_grid}{values for gridsearch (vector of integer valued numerics)}

\item{s2_grid}{values for gridsearch (vector of integer valued numerics)}

\item{k}{number of k-folds (integer greater than 1)}

\item{reps}{number of replications (positive integer)}

\item{B_type}{B-step descent type (string: 'gd')}

\item{Omega_type}{Omega-step descent type (string: 'gd' or 'min')}

\item{eta1}{B-step learning rate (positive numeric)}

\item{eta2}{Omega-step learning rate (positive numeric)}

\item{rho1}{B-step linesearch convergence parameter (positive numeric)}

\item{rho2}{Omega-step linesearch convergence parameter  (positive numeric)}

\item{beta1}{B-step linesearch shrinkage parameter (positive numeric)}

\item{beta2}{Omega-step linesearch shrinkage parameter (positive numeric)}

\item{epsilon}{convergence parameter (positive numeric)}

\item{max_iter}{maximum number of allowed iterations (positive integer)}

\item{suppress}{suppress maximum iteration warning (bool)}

\item{quiet}{quiet mode (bool)}

\item{seed}{set random seed (integer)}
}
\value{
A list containing the \code{mean} and \code{sd} of the
error over the replicates as well as the means and standard
deviations of the errors across each fold.
See also \code{\link{squared_error}}, \code{\link{k_folds}},
\code{\link{tsmvr_cv}}, and \code{\link{tsmvr_solve}}.
}
\description{
Calculates the mean and standard deviation of fold-averaged errors
derived from replicated tsmvr k-fold cross-validation experiments.
The error for each base model is the normalized squared
error between the true response and the predicted response
on a given cross-validation set.
}
