% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmvr_replicate.R
\name{tsmvr_replicate}
\alias{tsmvr_replicate}
\title{Replicated k-fold cross-validation for tsmvr}
\usage{
tsmvr_replicate(X, Y, s1, s2, pars, quiet = F, seed = NULL)
}
\arguments{
\item{X}{design matrix (n-by-p)}

\item{Y}{response matrix (n-by-q)}

\item{s1}{regressor matrix sparsity (positive integer)}

\item{s2}{covariance matrix sparsity (positive integer)}

\item{pars}{list of algorithm parameters; output of \code{set_parameters}}

\item{seed}{set random seed (integer)}

\item{k}{number of k-folds (integer greater than 1)}

\item{reps}{number of replications (positive integer)}
}
\value{
A list containing the \code{mean} and \code{sd} of the
error over the replicates as well as the means and standard
deviations of the errors across each fold.
}
\description{
Calculates the mean and standard deviation of fold-averaged errors
derived from replicated tsmvr k-fold cross-validation experiments.
The error for each base model is the normalized squared
error between the true response and the predicted response
on a given cross-validation set.
}
\note{
See also \code{\link{squared_error}},
\code{\link{k_folds}},
\code{\link{tsmvr_solve}},
\code{\link{tsmvr_cv}}, and
\code{\link{set_parameters}}.
}
